[Test]Move test to test module
